feat: finish base screen

+++ invman_flutter/lib/features/withdrawal/fee/components/form_component.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:go_router/go_router.dart';
 import 'package:invman_flutter/config/generated/l10n.dart';
 import 'package:invman_flutter/core/components/components.dart';
+import 'package:invman_flutter/core/navigation/navigation.dart';
 import 'package:invman_flutter/core/providers/providers.dart';
 import 'package:invman_flutter/core/utils/utils.dart';
 import 'package:invman_flutter/features/withdrawal/fee/fee.dart';
@@ -19,7 +19,7 @@ class WithdrawalFeeFormComponent extends ConsumerWidget {
 
     return BaseStateComponent(
       state: state,
-      onRefresh: () => provider.load(),
+      onErrorRefresh: () => provider.load(),
       successBuilder: (data) => Form(
         key: provider.formKey,
         child: SingleChildScrollView(
@@ -56,19 +56,20 @@ class WithdrawalFeeFormComponent extends ConsumerWidget {
                   final (success, message) = await provider.submit();
                   ToastUtils.message(message, success: success);
                   if (success) {
-                    context.pop();
+                    router.pop();
                   }
                 },
               ),
-              DeleteButton(
-                onPressed: () async {
-                  final (success, message) = await provider.delete();
-                  ToastUtils.message(message, success: success);
-                  if (success) {
-                    context.pop();
-                  }
-                },
-              )
+              if (id != 0)
+                DeleteButton(
+                  onPressed: () async {
+                    final (success, message) = await provider.delete();
+                    ToastUtils.message(message, success: success);
+                    if (success) {
+                      router.pop();
+                    }
+                  },
+                )
             ],
           ),
         ),
